customer_gateway/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
snowflake-connector-python
httpx
python-dotenv
